+++ src/entities/pools/fx/fxFactory.ts
@@ -1,5 +1,5 @@
 import { BasePool, BasePoolFactory } from '../';
-import { FxPool } from './';
+import { FxPool } from './fxPool';
 import { RawFxPool, RawPool } from '../../../data/types';
 
 export class FxPoolFactory implements BasePoolFactory {
