+++ src/entities/pools/fx/index.ts
@@ -1,3 +1,5 @@
 export * from './fxFactory';
 export * from './fxPool';
 export * from './fxMath';
+export * from './types';
+export * from './fxPoolToken';
