+++ src/entities/pools/fx/types.ts
@@ -1,0 +1,17 @@
+import { SwapKind } from '../../../types';
+import { FxPoolToken } from './fxPoolToken';
+
+export type FxPoolPairData = {
+    tIn: FxPoolToken;
+    tOut: FxPoolToken;
+    alpha: bigint;
+    beta: bigint;
+    delta: bigint;
+    lambda: bigint;
+    _oGLiq: bigint;
+    _nGLiq: bigint;
+    _oBals: bigint[];
+    _nBals: bigint[];
+    givenToken: FxPoolToken;
+    swapKind: SwapKind;
+};
