+++ src/entities/pools/gyroE/index.ts
@@ -1,3 +1,4 @@
 export * from './gyroEFactory';
 export * from './gyroEPool';
 export * from './gyroEMath';
+export * from './types';
