+++ src/entities/pools/linear/index.ts
@@ -1,3 +1,4 @@
 export * from './linearFactory';
 export * from './linearPool';
 export * from './linearMath';
+export * from './types';
