+++ src/entities/pools/linear/types.ts
@@ -1,0 +1,6 @@
+export type Params = {
+    fee: bigint;
+    rate: bigint;
+    lowerTarget: bigint;
+    upperTarget: bigint;
+};
